feat(gravsearch): Add feature toggle for topological sort optimisation.
