+++ webapi/src/main/resources/application.conf
@@ -292,6 +292,21 @@ app {
                 "Benjamin Geer <benjamin.geer@dasch.swiss>"
             ]
         }
+
+        gravsearch-dependency-optimisation {
+            description = "Optimise Gravsearch queries by reordering query patterns according to their dependencies."
+
+            available-versions = [ 1 ]
+            default-version = 1
+            enabled-by-default = yes
+            override-allowed = yes
+            expiration-date = "2021-12-01T00:00:00Z"
+
+            developer-emails = [
+                "Sepideh Alassi <sepideh.alassi@dasch.swiss>"
+                "Benjamin Geer <benjamin.geer@dasch.swiss>"
+            ]
+        }
     }
 
     shacl {
