+++ webapi/src/test/scala/org/knora/webapi/messages/util/search/gravsearch/prequery/NonTriplestoreSpecificGravsearchToCountPrequeryTransformerSpec.scala
@@ -2,6 +2,7 @@ package org.knora.webapi.util.search.gravsearch.prequery
 
 import org.knora.webapi.CoreSpec
 import org.knora.webapi.exceptions.AssertionException
+import org.knora.webapi.feature.FeatureFactoryConfig
 import org.knora.webapi.messages.IriConversions._
 import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
@@ -13,7 +14,6 @@ import org.knora.webapi.messages.util.search.gravsearch.types.{
   GravsearchTypeInspectionUtil
 }
 import org.knora.webapi.messages.util.search.gravsearch.{GravsearchParser, GravsearchQueryChecker}
-import org.knora.webapi.settings.KnoraSettingsImpl
 import org.knora.webapi.sharedtestdata.SharedTestDataADM
 
 import scala.collection.mutable.ArrayBuffer
@@ -26,7 +26,9 @@ private object CountQueryHandler {
 
   val anythingUser: UserADM = SharedTestDataADM.anythingAdminUser
 
-  def transformQuery(query: String, responderData: ResponderData, settings: KnoraSettingsImpl): SelectQuery = {
+  def transformQuery(query: String,
+                     responderData: ResponderData,
+                     featureFactoryConfig: FeatureFactoryConfig): SelectQuery = {
 
     val constructQuery = GravsearchParser.parseQuery(query)
 
@@ -51,7 +53,8 @@ private object CountQueryHandler {
       new NonTriplestoreSpecificGravsearchToCountPrequeryTransformer(
         constructClause = constructQuery.constructClause,
         typeInspectionResult = typeInspectionResult,
-        querySchema = constructQuery.querySchema.getOrElse(throw AssertionException(s"WhereClause has no querySchema"))
+        querySchema = constructQuery.querySchema.getOrElse(throw AssertionException(s"WhereClause has no querySchema")),
+        featureFactoryConfig = featureFactoryConfig
       )
 
     val nonTriplestoreSpecficPrequery: SelectQuery = QueryTraverser.transformConstructToSelect(
@@ -318,7 +321,9 @@ class NonTriplestoreSpecificGravsearchToCountPrequeryTransformerSpec extends Cor
     "transform an input query with a decimal as an optional sort criterion and a filter" in {
 
       val transformedQuery =
-        CountQueryHandler.transformQuery(inputQueryWithDecimalOptionalSortCriterionAndFilter, responderData, settings)
+        CountQueryHandler.transformQuery(inputQueryWithDecimalOptionalSortCriterionAndFilter,
+                                         responderData,
+                                         defaultFeatureFactoryConfig)
 
       assert(transformedQuery === transformedQueryWithDecimalOptionalSortCriterionAndFilter)
 
@@ -329,7 +334,7 @@ class NonTriplestoreSpecificGravsearchToCountPrequeryTransformerSpec extends Cor
       val transformedQuery =
         CountQueryHandler.transformQuery(inputQueryWithDecimalOptionalSortCriterionAndFilterComplex,
                                          responderData,
-                                         settings)
+                                         defaultFeatureFactoryConfig)
 
       assert(transformedQuery === transformedQueryWithDecimalOptionalSortCriterionAndFilterComplex)
